+++ Pomnesh/Pomnesh.Application/Services/ChatContextService.cs
@@ -1,7 +1,9 @@
-﻿using Pomnesh.Application.Dto;
+﻿using Pomnesh.API.Dto;
+using Pomnesh.Application.Dto;
 using Pomnesh.Application.DTO;
 using Pomnesh.Application.Interfaces;
 using Pomnesh.Domain.Entity;
+using Pomnesh.Application.Exceptions;
 using Pomnesh.Infrastructure.Interfaces;
 
 namespace Pomnesh.Application.Services;
@@ -20,32 +22,62 @@ public class ChatContextService(IBaseRepository<ChatContext> contextRepository) 
         return await contextRepository.Add(context);
     }
 
-    public async Task<ChatContext?> Get(long id)
+    public async Task<ChatContextResponse?> Get(long id)
     {
         var result = await contextRepository.GetById(id);
-        return result;
+        if (result == null)
+            throw new ContextNotFoundError(id);
+        
+        return new ChatContextResponse
+        {
+            Id = result.Id,
+            MessageId = result.MessageId,
+            MessageText = result.MessageText
+        };
     }
     
-    public async Task<IEnumerable<ChatContext>> GetAll()
+    public async Task<IEnumerable<ChatContextResponse>> GetAll()
     {
         var result = await contextRepository.GetAll();
-        return result;
+        
+        var chatContextResponse = new List<ChatContextResponse>();
+        foreach (var chatContext in result)
+        {
+            var responseDto = new ChatContextResponse
+            {
+                Id = chatContext.Id,
+                MessageId = chatContext.MessageId,
+                MessageText = chatContext.MessageText,
+                MessageDate = chatContext.MessageDate
+            };
+            chatContextResponse.Add(responseDto);
+        }
+
+        return chatContextResponse;
     }
 
     public async Task Update(ChatContextUpdateDto data)
     {
-        var context = new ChatContext
+        var chatContext = await contextRepository.GetById(data.Id);
+        if (chatContext == null)
+            throw new ContextNotFoundError(data.Id);
+
+        var updatedChatContext = new ChatContext
         {
             Id = data.Id,
             MessageId = data.MessageId,
             MessageText = data.MessageText,
             MessageDate = data.MessageDate,
         };
-        await contextRepository.Update(context);
+        await contextRepository.Update(updatedChatContext);
     }
 
     public async Task Delete(long id)
     {
+        var chatContext = await contextRepository.GetById(id);
+        if (chatContext == null)
+            throw new ContextNotFoundError(id);
+        
         await contextRepository.Delete(id);
     }
 }
